+++ asyncband/src/mutex/mod.rs
@@ -279,7 +279,7 @@ unsafe impl<T: ?Sized + Send + Sync> Sync for MutexGuard<'_, T> {}
 
 impl<T: ?Sized> Drop for MutexGuard<'_, T> {
     fn drop(&mut self) {
-        self.lock.s.release(1);
+        self.lock.s.release_all_held(1);
     }
 }
 
@@ -444,7 +444,7 @@ unsafe impl<T: ?Sized + Send + Sync> Sync for OwnedMutexGuard<T> {}
 
 impl<T: ?Sized> Drop for OwnedMutexGuard<T> {
     fn drop(&mut self) {
-        self.lock.s.release(1);
+        self.lock.s.release_all_held(1);
     }
 }
 
@@ -639,7 +639,7 @@ unsafe impl<T: ?Sized + Sync> Sync for MappedMutexGuard<'_, T> {}
 
 impl<T: ?Sized> Drop for MappedMutexGuard<'_, T> {
     fn drop(&mut self) {
-        self.s.release(1);
+        self.s.release_all_held(1);
     }
 }
 
@@ -845,7 +845,7 @@ unsafe impl<T: ?Sized + Send + Sync, U: ?Sized + Send + Sync> Sync for OwnedMapp
 impl<T: ?Sized, U: ?Sized> Drop for OwnedMappedMutexGuard<T, U> {
     fn drop(&mut self) {
         // Release the lock by calling release on the semaphore
-        self.lock.s.release(1);
+        self.lock.s.release_all_held(1);
     }
 }